- Services now implement `Process`. - Mods are now services. - Refactored some packages. - Started on a GUI system
See more in pull request
