+++ src/main/java/cc/hyperium/mixins/client/resources/MixinLocale.java
@@ -1,6 +1,6 @@
 package cc.hyperium.mixins.client.resources;
 
-import cc.hyperium.services.i18n.HyperiumLocale;
+import cc.hyperium.processes.services.i18n.HyperiumLocale;
 import net.minecraft.client.resources.Locale;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
