+++ src/main/kotlin/cc/hyperium/game/guis/EmptyGui.kt
@@ -1,6 +1,6 @@
 package cc.hyperium.game.guis
 
-import cc.hyperium.services.keybind.LWJGLKey
+import cc.hyperium.processes.services.keybind.LWJGLKey
 import net.minecraft.client.gui.GuiScreen
 
 open class EmptyGui(val keybind: LWJGLKey) : GuiScreen() {
