+++ src/main/kotlin/cc/hyperium/game/mods/ToggleSprint.kt
@@ -1,0 +1,6 @@
+package cc.hyperium.game.mods
+
+import cc.hyperium.processes.services.game.mod.Mod
+
+object ToggleSprint : Mod() {
+}
