+++ src/main/kotlin/cc/hyperium/game/mods/visual/Fullbright.kt
@@ -1,0 +1,6 @@
+package cc.hyperium.game.mods.visual
+
+import cc.hyperium.processes.services.game.mod.Mod
+
+object Fullbright : Mod() {
+}
