+++ src/main/kotlin/cc/hyperium/processes/services/keybind/Bindable.kt
@@ -1,0 +1,3 @@
+package cc.hyperium.processes.services.keybind
+
+interface Bindable<Key : KeyboardKey> : KeyListener<Key>
