+++ src/main/kotlin/cc/hyperium/processes/services/keybind/KeyListener.kt
@@ -1,0 +1,16 @@
+package cc.hyperium.processes.services.keybind
+
+interface KeyListener<Key : KeyboardKey> {
+
+    fun onKeyPress(key: Key)
+
+    fun onKeyRelease(key: Key)
+
+}
+
+abstract class AbstractKeyListener<Key : KeyboardKey> : KeyListener<Key> {
+
+    override fun onKeyPress(key: Key) {}
+
+    override fun onKeyRelease(key: Key) {}
+}
