--- src/main/kotlin/cc/hyperium/services/keybind/Bindable.kt
@@ -1,3 +1,0 @@
-package cc.hyperium.services.keybind
-
-interface Bindable<Key : KeyboardKey> : KeyListener<Key>
