--- src/main/kotlin/cc/hyperium/services/keybind/KeyListener.kt
@@ -1,16 +1,0 @@
-package cc.hyperium.services.keybind
-
-interface KeyListener<Key : KeyboardKey> {
-
-    fun onKeyPress(key: Key)
-
-    fun onKeyRelease(key: Key)
-
-}
-
-abstract class AbstractKeyListener<Key : KeyboardKey> : KeyListener<Key> {
-
-    override fun onKeyPress(key: Key) {}
-
-    override fun onKeyRelease(key: Key) {}
-}
